refactor(macos): split terminal runtime service owners
